+++ src/blocks/lightbox/supports.ts
@@ -1,0 +1,9 @@
+import type { BlockSupports } from "@atomicsmash/blocks-helpers";
+
+export const supports = {
+	color: {
+		background: true,
+		text: false,
+	},
+} as const satisfies BlockSupports;
+export type Supports = typeof supports;